fixed css of add match result

+++ tourni-management/src/main/java/com/tournament/management/service/TourniManagementService.java
@@ -13,5 +13,5 @@ public interface TourniManagementService {
 
     List<TournamentDTO> getAllTournaments();
 
-    List<TeamDTO> getAllTeamsByTournamentId(int tournamentId);
+    List<TeamDTO> getAllTeamsByTournamentId(Long tournamentId);
 }
